Moved Footer and Navbar into the "Globals" folder for better clarity and access. Also updated dependancies to resolve bugs
