+++ client/src/pages/GetInvolved.jsx
@@ -1,7 +1,7 @@
-import Navbar from "../components/home/layout/Navbar";
+import Navbar from "../components/globals/Navbar";
 import GetInvolvedHeader from "../components/getInvolved/GetInvolvedHeader.jsx"; 
 import DPVCards from "../components/getInvolved/DPVCards";
-import Footer from "../components/home/layout/Footer"
+import Footer from "../components/globals/Footer"
 
 export default function GetInvolved() {
     return (
