+++ client/src/pages/IndividualProgram.jsx
@@ -1,4 +1,4 @@
-import Navbar from "../components/home/layout/Navbar";
+import Navbar from "../components/globals/Navbar";
 import Header from "../components/individualprogram/Header";
 import Context from "../components/individualprogram/Context";
 import Statistics from "../components/individualprogram/Statistics";
