+++ client/src/pages/Partners.jsx
@@ -1,4 +1,4 @@
-import Navbar from "../components/home/layout/Navbar";
+import Navbar from "../components/globals/Navbar";
 import OurPartners from "../components/about/layout/OurPartners.jsx";
 
 export default function Partners(){
