+++ client/src/pages/Programs.jsx
@@ -1,7 +1,7 @@
-import Navbar from "../components/home/layout/Navbar";
+import Navbar from "../components/globals/Navbar";
 import OurPrograms from "../components/allPrograms/OurPrograms";
 import CareAndFair from "../components/allPrograms/CareAndFair";
-import Footer from "../components/home/layout/Footer"
+import Footer from "../components/globals/Footer"
 
 export default function Programs() {
     return (
