+++ client/src/pages/ReusableProgram.jsx
@@ -1,6 +1,6 @@
-import Navbar from "../components/home/layout/Navbar";
+import Navbar from "../components/globals/Navbar";
 import ReusableProgramPage from "../components/reusableprogram/layout/ReusableProgramPage";
-import Footer from "../components/home/layout/Footer"
+import Footer from "../components/globals/Footer"
 
 export default function ReusableProgram() {
     return (
