+++ client/src/pages/Team.jsx
@@ -1,4 +1,4 @@
-import Navbar from "../components/home/layout/Navbar";
+import Navbar from "../components/globals/Navbar";
 import OurTeam from "../components/team/OurTeam";
 
 export default function Team () {
